Move view mixin tests to their own source file
